requirements.txt
Django==3.0.6
django-webpack-loader
djangorestframework
yapf